workers/requirements.txt
supabase==2.9.0
yfinance==0.2.54
pandas==2.2.2
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.2
tenacity==8.5.0
structlog==24.4.0
pandas-ta==0.3.14b
anthropic==0.34.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
feedparser==6.0.11
beautifulsoup4==4.12.3
pdfplumber==0.11.4
